+++ ui/pages/DataManagement/components/SampleQc.jsx
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 
-import { validators } from 'shared/components/form/ReduxFormWrapper'
+import { validators } from 'shared/components/form/FormHelpers'
 
 import { getQcUploadStats } from '../selectors'
 import { uploadQcPipelineOutput } from '../reducers'
