+++ ui/pages/Login/components/ForgotPassword.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
-import { validators } from 'shared/components/form/ReduxFormWrapper'
+import { validators } from 'shared/components/form/FormHelpers'
 import { forgotPassword } from '../reducers'
 import UserFormLayout from './UserFormLayout'
