+++ ui/pages/Project/components/ProjectCollaborators.jsx
@@ -12,7 +12,7 @@ import { HorizontalSpacer } from 'shared/components/Spacers'
 import DeleteButton from 'shared/components/buttons/DeleteButton'
 import UpdateButton from 'shared/components/buttons/UpdateButton'
 import { RadioGroup, AddableSelect } from 'shared/components/form/Inputs'
-import { validators } from 'shared/components/form/ReduxFormWrapper'
+import { validators } from 'shared/components/form/FormHelpers'
 import { USER_NAME_FIELDS } from 'shared/utils/constants'
 
 import { updateCollaborator } from '../reducers'
