+++ ui/pages/Report/components/SampleMetadata.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
-import ReduxFormWrapper from 'shared/components/form/ReduxFormWrapper'
+import FormWrapper from 'shared/components/form/FormWrapper'
 import { BaseSemanticInput } from 'shared/components/form/Inputs'
 import { loadSampleMetadata } from '../reducers'
 import { getSampleMetadataLoading, getSampleMetadataLoadingError, getSampleMetadataRows, getSampleMetadataColumns } from '../selectors'
@@ -25,7 +25,7 @@ const FIELDS = [
 ]
 
 const SampleMetadataFilters = React.memo(({ load }) => (
-  <ReduxFormWrapper
+  <FormWrapper
     onSubmit={load}
     fields={FIELDS}
     noModal
