+++ ui/pages/Search/components/VariantSearchResults.jsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 
 import { loadSearchedVariants } from 'redux/rootReducer'
 import { InlineToggle } from 'shared/components/form/Inputs'
-import { helpLabel, StyledForm } from 'shared/components/form/ReduxFormWrapper'
+import { helpLabel, StyledForm } from 'shared/components/form/FormHelpers'
 import VariantSearchResults, { DisplayVariants } from 'shared/components/panel/search/VariantSearchResults'
 
 import { updateCompoundHetDisplay, loadSingleSearchedVariant, loadProjectFamiliesContext } from '../reducers'
