+++ ui/shared/components/panel/view-fields/BaseFieldView.jsx
@@ -10,7 +10,7 @@ import DeleteButton from '../../buttons/DeleteButton'
 import UpdateButton from '../../buttons/UpdateButton'
 import { HorizontalSpacer } from '../../Spacers'
 import { ButtonLink } from '../../StyledComponents'
-import ReduxFormWrapper from '../../form/FormHelpers'
+import FormWrapper from '../../form/FormWrapper'
 
 const FieldValue = styled.div`
   padding-bottom: ${props => (props.compact ? 0 : '15px')}; 
@@ -105,7 +105,7 @@ class BaseFieldView extends React.PureComponent {
             />
           ) : (
             <Segment compact>
-              <ReduxFormWrapper
+              <FormWrapper
                 noModal
                 inline
                 key="edit"
